refactor(app): move the remaining media UI into fork-owned files
Four extractions, all the same shape: the media logic moves to a path upstream does
not have, and what stays behind in upstream's file is a call.

PlaygroundChatTemplate keeps a hook call and one element instead of thirty-two lines
of state and a role-checked pair of buttons: useMessageMedia holds the four lists
and their setters, MessageMediaButtons holds the toolbar contribution, and
PlaygroundMessageMedia now takes the hook's own state type rather than eight props.
That file goes from 97 added lines to 52.

ChatTemplateMessageCard loses the four media part renderers, the variable pill and
their shared container wrapper — 150 lines — to components/prompt/media/. Its titles
are spread in from a small module both it and the media components can import, so
neither has to import the other and there is no cycle. Exporting
ChatTemplateMessagePartContainer is the one word upstream's file gives up.

PromptChatMessagesCard's four sequential part checks become one:
ChatTemplateMessageMediaPart picks the renderer, so adding a media kind happens
there rather than in upstream's dispatch.

schemas.ts and SpanImage.tsx follow the same pattern — a spread of
mediaMessageShape, and a MediaNotAnImage component — taking 33 and 40 added lines
down to 2 and 17.

Across this and the two changes before it, the frontend's footprint in files
upstream owns falls from 824 added lines to 401, and its churn-weighted risk from
2260 to 877.
