+++ app/src/components/core/types/style.ts
@@ -189,15 +189,6 @@ export type ColorValue =
   | "chartreuse-1200"
   | "chartreuse-1300"
   | "chartreuse-1400"
-  | "orange-100"
-  | "orange-200"
-  | "orange-300"
-  | "orange-400"
-  | "orange-500"
-  | "orange-600"
-  | "orange-700"
-  | "orange-800"
-  | "orange-900"
   | "orange-1000"
   | "orange-1100"
   | "orange-1200"
